refactro profile visualizations
