+++ pypesto/profile/profile.py
@@ -1,14 +1,15 @@
 import logging
 import numpy as np
-from typing import Any, Callable, Dict
+from typing import Callable
 
 from ..objective.constants import GRAD
 from ..optimize import Optimizer
 from ..problem import Problem
-from ..result import Result, ProfileResult
+from ..result import Result
 from .result import ProfilerResult
 from .profile_next_guess import next_guess
 from .options import ProfileOptions
+from .util import initialize_profile
 
 logger = logging.getLogger(__name__)
 
@@ -39,15 +40,15 @@ def parameter_profile(
         The optimizer to be used along each profile.
     profile_index:
         array with parameter indices, whether a profile should
-        be computed (1) or not (0)
-        Default is all profiles should be computed
+        be computed (1) or not (0).
+        Default is all profiles should be computed.
     profile_list:
         integer which specifies whether a call to the profiler should create
         a new list of profiles (default) or should be added to a specific
-        profile list
+        profile list.
     result_index:
         index from which optimization result profiling should be started
-        (default: global optimum, i.e., index = 0)
+        (default: global optimum, i.e., index = 0).
     next_guess_method:
         function handle to a method that creates the next starting point for
         optimization in profiling.
@@ -59,7 +60,6 @@ def parameter_profile(
     result:
         The profile results are filled into `result.profile_result`.
     """
-
     # Handling defaults
     # profiling indices
     if profile_index is None:
@@ -221,138 +221,3 @@ def walk_along_profile(
     problem.unfix_parameters(i_par)
 
     return current_profile
-
-
-def initialize_profile(
-        problem: Problem,
-        result: Result,
-        result_index: int,
-        profile_index: np.ndarray,
-        profile_list: int
-) -> float:
-    """
-    This function initializes profiling based on a previous optimization.
-
-    Parameters
-    ----------
-    problem:
-        The problem to be solved.
-    result:
-        A result object to initialize profiling and to append the profiling
-        results to. For example, one might append more profiling runs to a
-        previous profile, in order to merge these.
-        The existence of an optimization result is obligatory.
-    result_index:
-        index from which optimization result profiling should be started
-    profile_index:
-        array with parameter indices, whether a profile should
-        be computed (1) or not (0)
-        Default is all profiles should be computed
-    profile_list:
-        integer which specifies whether a call to the profiler should create
-        a new list of profiles (default) or should be added to a specific
-        profile list
-
-    Returns
-    -------
-    global_opt:
-        log-posterior at global optimum.
-    """
-    # Check whether an optimization result is existing
-    if result.optimize_result is None:
-        raise ValueError(
-            "Optimization has to be carried out before profiling can be done.")
-
-    tmp_optimize_result = result.optimize_result.as_list()
-
-    # Check if new profile_list is to be created
-    if profile_list is None:
-        result.profile_result.append_empty_profile_list()
-
-    # get the log-posterior of the global optimum
-    global_opt = tmp_optimize_result[0]["fval"]
-
-    # fill the list with optimization results where necessary
-    fill_profile_list(
-        profile_result=result.profile_result,
-        optimizer_result=tmp_optimize_result[result_index],
-        profile_index=profile_index,
-        profile_list=profile_list,
-        problem_dimension=problem.dim_full,
-        global_opt=global_opt)
-
-    # return the log-posterior of the global optimum (needed in order to
-    # compute the log-posterior-ratio)
-    return global_opt
-
-
-def fill_profile_list(
-        profile_result: ProfileResult,
-        optimizer_result: Dict[str, Any],
-        profile_index: np.ndarray,
-        profile_list: int,
-        problem_dimension: int,
-        global_opt: float
-) -> None:
-    """
-    This is a helper function for initialize_profile
-
-    Parameters
-    ----------
-    profile_result:
-        A list of profiler result objects.
-    optimizer_result:
-        A local optimization result.
-    profile_index:
-        array with parameter indices, whether a profile should
-        be computed (1) or not (0).
-        Default is all profiles should be computed.
-    profile_list:
-        integer which specifies whether a call to the profiler should
-        create a new list of profiles (default) or should be added to a
-        specific profile list.
-    problem_dimension:
-        number of parameters in the unreduced problem.
-    global_opt:
-        log-posterior at global optimum.
-    """
-
-    if optimizer_result[GRAD] is not None:
-        gradnorm = np.linalg.norm(optimizer_result[GRAD])
-    else:
-        gradnorm = None
-
-    # create blank profile
-    new_profile = ProfilerResult(
-        x_path=optimizer_result["x"],
-        fval_path=np.array([optimizer_result["fval"]]),
-        ratio_path=np.array([np.exp(global_opt - optimizer_result["fval"])]),
-        gradnorm_path=gradnorm,
-        exitflag_path=optimizer_result["exitflag"],
-        time_path=np.array([0.]),
-        time_total=0.,
-        n_fval=0,
-        n_grad=0,
-        n_hess=0,
-        message=None)
-
-    if profile_list is None:
-        # All profiles have to be created from scratch
-        for i_parameter in range(0, problem_dimension):
-            if profile_index[i_parameter] > 0:
-                # Should we create a profile for this index?
-                profile_result.append_profiler_result(new_profile)
-            else:
-                # if no profile should be computed for this parameter
-                profile_result.append_profiler_result(None)
-
-    else:
-        for i_parameter in range(0, problem_dimension):
-            # We append to an existing list
-            if profile_index[i_parameter] > 0:
-                # Do we have to create a new profile?
-                create_new = (profile_result.list[profile_list][i_parameter]
-                              is None) and (profile_index[i_parameter] > 0)
-                if create_new:
-                    profile_result.set_profiler_result(
-                        new_profile, i_parameter)
